Moved logic participatn component to viewer service
This allows state persistance between drawer views/routes
